+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## next
 
+- Add the description to the `hook event-types` view.
+
 ## 0.0.15
 
 - Ability to list studies in a different project.